+++ R/data.R
@@ -1,0 +1,14 @@
+#' idaifieldR Demodata
+#'
+#' Test and demonstration data to illustrate the packages functionality, has
+#' no real world representation.
+#'
+#' @format ## `idaifieldr_demodata`
+#' A list (idaifield_docs) with 108 elements:
+#' \describe{
+#'   \item{identifier}{identifier / Kurzbezeichner}
+#'   \item{all other values}{entries in the field database}
+#'   ...
+#' }
+#' @source Random collection of imagined data
+"idaifieldr_demodata"