+++ R/get_uid_list.R
@@ -1,40 +1,54 @@
-#' get_uid_list
+#' get_uid_list: Get the index of an idaifield_docs/resources object.
 #'
-#' All resources from `idaifield` are referenced with their UID in the relations
-#' fields. Therefore, for many purposes a lookup-table needs to be provided
-#' in order to get to the actual identifiers and names of the resources
-#' referenced.
+#' All resources in the project databases in iDAI.field / Field Desktop are
+#' stored and referenced with their Universally Unique Identifier (UUID)
+#' in the relations fields. Therefore, for many purposes a lookup-table needs
+#' to be provided in order to get to the actual identifiers of the resources
+#' referenced. Single UUIDs or vectors of UUIDs can be replaced individually
+#' using `replace_uid()` from this package.
 #'
 #' This function is also good for a quick overview / a list of all the
 #' resources that exist along with their identifiers and short descriptions
 #' and can be used to select the resources along their respective Types
 #' (e.g. Pottery, Layer etc.). Please note that in any case the internal
 #' names of everything will be used. If you relabeled `Trench` to `Schnitt` in
-#' your language-configuration, this will still be `Trench` here.
-#' None of these functions have any respect for language configuration!
+#' your language-configuration, the name will still be `Trench` here.
+#' None of these functions have any respect for language settings of a
+#' project configuration, i.e. the front end languages of valuelists and fields
+#' are not displayed, and instead their background names are used. You can see
+#' these in the project configuration settings.
 #'
-#' @param idaifield_docs An object as returned by get_idaifield_docs
-#' @param verbose TRUE or FALSE (= anything but TRUE), TRUE returns a list
-#' including identifier and shorttitle which is more convenient for humans,
-#' and FALSE returns only UID and type, which is enough for internal selections
+#' @param idaifield_docs An object as returned by `get_idaifield_docs()`
+#' @param verbose TRUE or FALSE. Defaults to FALSE. TRUE returns a list
+#' including identifier and shortDescription which is more convenient to read,
+#' and FALSE returns only UUID, type and basic relations, which is sufficient
+#' for internal use.
 #' @param gather_trenches defaults to FALSE. If TRUE, adds another column that
 #' records the Place each corresponding Trench and its sub-resources lie within.
-#' (Useful for grouping the finds of several trenches.)
+#' (Useful for grouping the finds of several trenches, but will only work if the
+#' project database is organized accordingly.)
+#' @param language the short name (e.g. "en", "de", "fr") of the language that
+#' is preferred for the multi-language input "shortDescription",
+#' defaults to english ("en") and will select other available languages in
+#' alphabetical order if the selected language is not available.
 #'
-#' @return a list of UIDs and their Types, Identifiers and shortDescriptions
+#' @return a data.frame with identifiers and corresponding UUIDs along with
+#' the type, basic relations and depending on settings place and
+#' shortDescription of each element
 #' @export
 #'
 #' @examples
 #' \dontrun{
 #' connection <- connect_idaifield(serverip = "127.0.0.1",
-#' user = "R", pwd = "hallo")
+#'                                 user = "R", pwd = "hallo")
 #' idaifield_docs <- get_idaifield_docs(connection = connection,
-#' projectname = "rtest", simplified = FALSE)
+#'                                      projectname = "rtest")
 #'
 #' uidlist <- get_uid_list(idaifield_docs, verbose = TRUE)
 #' }
 get_uid_list <- function(idaifield_docs, verbose = FALSE,
-                         gather_trenches = FALSE) {
+                         gather_trenches = FALSE,
+                         language = "en") {
   idaifield_docs <- check_and_unnest(idaifield_docs)
 
   ncol <- 5
@@ -44,6 +58,7 @@ get_uid_list <- function(idaifield_docs, verbose = FALSE,
     ncol <- 7
     colnames <- c(colnames, "shortDescription", "liesWithinLayer")
   }
+
 
   uidlist <- data.frame(matrix(nrow = length(idaifield_docs), ncol = ncol))
   colnames(uidlist) <- colnames
@@ -74,10 +89,10 @@ get_uid_list <- function(idaifield_docs, verbose = FALSE,
                                         function(x) na_if_empty(x$relation.liesWithin)))
   }
 
-
   if (verbose) {
-    uidlist$shortDescription <- unlist(lapply(idaifield_docs,
-                                              function(x) na_if_empty(x$shortDescription)))
+    # get all entries for shortDescription
+    desc <- lapply(idaifield_docs, function(x) na_if_empty(x$shortDescription))
+    uidlist$shortDescription <- gather_languages(desc, language = language)
     uidlist$liesWithinLayer <- unlist(lapply(idaifield_docs,
                                              function(x) na_if_empty(x$relation.liesWithinLayer)))
   }