+++ R/unnest_docs.R
@@ -7,14 +7,9 @@
 #' about who created and modified the resource at what time and
 #' is irrelevant for any analysis of the database contents itself.
 #'
-#' @param idaifield_docs A list as provided by `sofa::db_alldocs(...)` when
-#' importing from an iDAI.field 2 / Field Desktop-database
-#' using `get_idaifield_docs()`.
-#' `get_idaifield_docs()` employs this function already
-#' when setting `simplified = TRUE`, which is the default.
-#' Mostly there is no need to deal with `unnest_docs()`. If one chooses
-#' `get_idaifield_docs(..., simplified  = FALSE)`, it is possible to use
-#' `unnest_docs()` on the resulting list to simplify it.
+#' @param idaifield_docs A list as provided by `get_idaifield_docs()`.
+#' `get_idaifield_docs()` employs this function already when
+#' setting `raw = FALSE`.
 #'
 #' @export
 #' @keywords internal
@@ -46,6 +41,8 @@ unnest_docs <- function(idaifield_docs) {
                                                     "connection")
     attr(idaifield_resources, "projectname") <- attr(idaifield_docs,
                                                      "projectname")
+    attr(idaifield_resources, "config") <- attr(idaifield_docs,
+                                                     "config")
     return(idaifield_resources)
   } else if (check_result["idaifield_resources"]) {
     message("The list was already unnested to resource-level.")