+++ R/utilities.R
@@ -120,8 +120,6 @@ check_for_sublist <- function(single_resource_field) {
   return(has_sublist)
 }
 
-
-
 #' check_if_uid
 #'
 #' @param string A character string or vector of character strings that should
@@ -174,4 +172,104 @@ reorder_colnames <- function(colnames, order = "default") {
   new_order <- colnames[new_order]
   colnames <- c(new_order, not_ordered)
   return(colnames)
+}
+
+
+
+#' Gather multilanguage fields
+#'
+#' @param input_list a list with character values containing (or not)
+#' sublists for each language
+#' @param language the short name (e.g. "en", "de", "fr") of the language that
+#' is preferred for the fields, defaults to english ("en")
+#' @param silant TRUE/FALSE: Should gather_languages()
+#' issue messages and warnings?
+#'
+#' @return a vector containing the values
+#'
+#' @keywords internal
+#'
+#' @examples
+#' \dontrun{
+#' input_list <- list(list("en" = "English text", "de" = "Deutscher Text"),
+#'                    list("en" = "Another english text", "de" = "Weiterer dt. Text"))
+#' gather_languages(input_list, language = "de")
+#' }
+gather_languages <- function(input_list, language = "en", silent = FALSE) {
+  # if this has a sublist / more than one entry, it means that there is
+  # more than one language
+  if (check_for_sublist(input_list)) {
+    # try to get the selected language or english
+    res <- lapply(input_list, function(x) na_if_empty(unlist(x[language])))
+    res <- unlist(res)
+    if (all(is.na(res))) {
+      if (!silent) {
+        # issue warning is this does not work
+        warning("No language selected or selected language not available, using first entry")
+      }
+      # instead get the first entry
+      res <- unlist(lapply(input_list, function(x) x[1]))
+    }
+    if (any(is.na(res))) {
+      # if there are still NA-values, assume that sometimes people only filled
+      # in one language and loop through all languages to fill the NA values
+      languages <- sort(unique(unlist(lapply(input_list,
+                                             function(x) names(x)))))
+      for (i in seq_along(languages)) {
+        res_sec <- lapply(input_list, function(x) na_if_empty(unlist(x[languages[i]])))
+        res_sec <- unlist(res_sec)
+        res <- ifelse(is.na(res), res_sec, res)
+        no_list_ind <- unlist(lapply(input_list, is.character))
+        no_list_ind <- which(no_list_ind)
+        res[no_list_ind] <- unlist(input_list)[no_list_ind]
+      }
+    }
+  } else {
+    res <- unlist(lapply(input_list, function(x) na_if_empty(x)))
+  }
+  # remove the names
+  res <- unname(res)
+  return(res)
+}
+
+#' Ping the Field-Database
+#'
+#' Helper for all functions that access the database to display
+#' specific warning messages.
+#'
+#' @param idaifield_connection The connection as returned by `connect_idaifield()`
+#' @param fail Should the function stop execution?
+#'
+#' @return returns no value
+#' @keywords internal
+#'
+#' @examples
+#' \dontrun{
+#' idaifield_connection <- connect_idaifield()
+#' idf_ping(idaifield_connection)
+#' }
+#'
+idf_ping <- function(idaifield_connection) {
+  ping <- try(sofa::ping(idaifield_connection), silent = TRUE)
+  if(inherits(ping, "try-error")) {
+    if(grepl("Connection refused", ping[1])) {
+      message <- paste0(ping[1], ":
+
+      Is Field Desktop / iDAI.field running?")
+    } else if (grepl("password", ping[1])) {
+      message <- paste0(ping[1], ":
+
+      Check if the password matches your password
+      in the settings of your Field Desktop Client and try again!")
+    } else if (grepl("invalid char", ping[1])) {
+      message <- paste0(ping[1], ":
+
+      Are you sure you used the correct version number?")
+    } else {
+      message <- ping[1]
+    }
+    return(message)
+  } else {
+    return(TRUE)
+  }
 }