+++ tests/testthat/test-get_idaifield_docs.R
@@ -1,42 +1,11 @@
+test_that("raw has connection as attribute", {
+  conn <- suppressWarnings(connect_idaifield(pwd = "wrongpwd"))
+  expect_error(get_idaifield_docs(conn))
+})
+
 skip_on_cran()
+connection <- skip_if_no_connection()
 
-connection <- connect_idaifield(serverip = "127.0.0.1",
-                                user = "R", pwd = "hallo")
-
-tryCatch({
-  sofa::ping(connection)
-},
-error = function(cond) {
-  skip("Test skipped, needs DB-connection")
-})
-
-
-test_that("works with 2", {
-  conn2 <- connect_idaifield(serverip = "127.0.0.1",
-                             user = "R", pwd = "hallo",
-                             version = 2)
-  expect_equal(conn2$port, 3000)
-})
-
-test_that("works with 3", {
-  conn2 <- connect_idaifield(serverip = "127.0.0.1",
-                             user = "R", pwd = "hallo",
-                             version = 3)
-  expect_equal(conn2$port, 3001)
-})
-
-test_that("works with '3'", {
-  conn2 <- connect_idaifield(serverip = "127.0.0.1",
-                             user = "R", pwd = "hallo",
-                             version = "3")
-  expect_equal(conn2$port, 3001)
-})
-
-test_that("error when using character", {
-  expect_error(connect_idaifield(serverip = "127.0.0.1",
-                                 user = "R", pwd = "hallo",
-                                 version = "kuchen"))
-})
 
 test_docs_raw <- get_idaifield_docs(projectname = "rtest",
                                 connection = connection,
@@ -46,17 +15,38 @@ test_resources <- get_idaifield_docs(projectname = "rtest",
                                 connection = connection,
                                 raw = FALSE)
 
-test_simple <- simplify_idaifield(test_docs)
+test_simple <- simplify_idaifield(test_resources)
 
 
-test_that("has connection as attribute", {
+test_that("raw has connection as attribute", {
+  test <- attr(test_docs_raw, "connection")
+  expect_true("Cushion" %in% class(test))
+})
+
+test_that("resources has connection as attribute", {
   test <- attr(test_resources, "connection")
   expect_true("Cushion" %in% class(test))
 })
 
-test_that("has connection as attribute", {
-  test <- attr(test_resources, "connection")
+test_that("simple has connection as attribute", {
+  test <- attr(test_simple, "connection")
   expect_true("Cushion" %in% class(test))
+})
+
+
+test_that("raw has config as attribute", {
+  test <- attr(test_docs_raw, "config")
+  expect_identical(test$identifier, "Configuration")
+})
+
+test_that("resources has config as attribute", {
+  test <- attr(test_resources, "config")
+  expect_identical(test$identifier, "Configuration")
+})
+
+test_that("simple has config as attribute", {
+  test <- attr(test_simple, "config")
+  expect_identical(test$identifier, "Configuration")
 })
 
 
@@ -83,7 +73,7 @@ test_that("returns json", {
   expect_true(jsonlite::validate(output))
 })
 
-test_that("attaches connection as attribute", {
+test_that("attaches working connection as attribute", {
   test_conn <- attr(test_docs_raw, "connection")
   pinglist <- sofa::ping(test_conn)
   expect_true(is.list(pinglist))