+++ tests/testthat/test-simplify_idaifield_resources.R
@@ -104,17 +104,71 @@ test_that("returns same object if already simple", {
 })
 
 
+# language support
+data("idaifieldr_demodata")
+
+test_that("notify for config", {
+  expect_warning(simplify_idaifield(idaifieldr_demodata,
+                                    uidlist = get_uid_list(idaifieldr_demodata),
+                                    keep_geometry = FALSE,
+                                    language = "en",
+                                    replace_uids = TRUE),
+                 "custom")
+})
+
+
+test_that("notify for language", {
+  lang <- "de"
+  expect_message(suppressWarnings(simplify_idaifield(idaifieldr_demodata,
+                                                     uidlist = get_uid_list(idaifieldr_demodata),
+                                                     keep_geometry = FALSE,
+                                                     language = lang,
+                                                     replace_uids = TRUE)),
+                 lang)
+})
+
+test_that("notify for unavailable language", {
+  lang <- "tr"
+  expect_message(suppressWarnings(simplify_idaifield(idaifieldr_demodata,
+                                                     uidlist = get_uid_list(idaifieldr_demodata),
+                                                     keep_geometry = FALSE,
+                                                     language = lang,
+                                                     replace_uids = TRUE)),
+                 lang)
+})
+
+
+test_that("return correct language", {
+  lang <- "de"
+  val <- idaifieldr_demodata[[1]]$doc$resource$shortDescription$de
+  test <- suppressWarnings(simplify_idaifield(idaifieldr_demodata,
+                                              uidlist = get_uid_list(idaifieldr_demodata),
+                                              keep_geometry = FALSE,
+                                              language = lang,
+                                              replace_uids = TRUE))
+  expect_identical(test[[1]]$shortDescription, val)
+})
+
+
+test_that("return all languages", {
+  lang <- "all"
+  expect_message(test <- suppressWarnings(
+    simplify_idaifield(idaifieldr_demodata,
+                       uidlist = get_uid_list(idaifieldr_demodata),
+                       keep_geometry = FALSE,
+                       language = lang,
+                       replace_uids = TRUE)),
+    "all languages")
+  expect_true(all(c("en", "de") %in% names(test[[1]]$shortDescription)))
+})
+
+
+
+## with db connection
+
 skip_on_cran()
 
-connection <- connect_idaifield(serverip = "127.0.0.1",
-                                user = "R", pwd = "hallo")
-
-tryCatch({
-  sofa::ping(connection)
-},
-error = function(cond) {
-  skip("Test skipped, needs DB-connection")
-})
+connection <- skip_if_no_connection()
 
 test_that("colnames from checkboxes are spread", {
   test_simple <- simplify_idaifield(test_resources,
@@ -124,3 +178,4 @@ test_that("colnames from checkboxes are spread", {
   namesvec <- namesvec[grepl("testAnkreuzfeld", namesvec)]
   expect_gt(length(namesvec), 1)
 })
+